Skip execution if logs from previous run exists

+++ vagrant_init_destroy_boxes.sh
@@ -74,15 +74,20 @@ main() {
       export VAGRANT_BOX_NAME_SHORT=`basename $VAGRANT_BOX_FILE | cut -d - -f 1,2,3`
       export VAGRANT_BOX_PROVIDER=${VAGRANT_BOX_NAME##*-}
       export VAGRANT_CWD="$TMPDIR/$VAGRANT_BOX_NAME_SHORT"
+      export LOG_FILE="$LOGDIR/${VAGRANT_BOX_NAME}_vagrant_init_destroy_boxes.log"
 
-      echo -e "*** ${VAGRANT_BOX_FILE} [$VAGRANT_BOX_NAME] ($VAGRANT_BOX_PROVIDER) ($TMPDIR/$VAGRANT_BOX_NAME_SHORT)" | tee $LOGDIR/$VAGRANT_BOX_NAME
+      if [ -f $LOG_FILE ]; then
+        echo -e "\n*** Logfile \"$LOG_FILE\" exist, please remove it... Skipping...\n"
+        continue
+      fi
+      echo -e "*** ${VAGRANT_BOX_FILE} [$VAGRANT_BOX_NAME] ($VAGRANT_BOX_PROVIDER) ($TMPDIR/$VAGRANT_BOX_NAME_SHORT)" | tee $LOG_FILE
       test -d "$VAGRANT_CWD" && rm -rf "$VAGRANT_CWD"
       mkdir "$VAGRANT_CWD"
 
       vagrant_box_add
       vagrant_init_up
 
-      check_vagrant_vm 2>&1 | tee -a $LOGDIR/$VAGRANT_BOX_NAME
+      check_vagrant_vm 2>&1 | tee -a $LOG_FILE
 
       #echo "Press ENTER to destroy the VMs"
       #read A
